move sort functions to where they are used
