+++ apollo-federation/src/operation/contains.rs
@@ -1,6 +1,4 @@
 use apollo_compiler::executable;
-use apollo_compiler::Name;
-use apollo_compiler::Node;
 
 use super::FieldSelection;
 use super::FragmentSpreadSelection;
@@ -8,120 +6,6 @@ use super::HasSelectionKey;
 use super::InlineFragmentSelection;
 use super::Selection;
 use super::SelectionSet;
-
-/// Sort an input value, which means specifically sorting their object values by keys (assuming no
-/// duplicates). This is used for hashing input values in a way consistent with [same_value()].
-fn sort_value(value: &mut executable::Value) {
-    use apollo_compiler::executable::Value;
-    match value {
-        Value::List(elems) => {
-            elems
-                .iter_mut()
-                .for_each(|value| sort_value(value.make_mut()));
-        }
-        Value::Object(pairs) => {
-            pairs
-                .iter_mut()
-                .for_each(|(_, value)| sort_value(value.make_mut()));
-            pairs.sort_by(|left, right| left.0.cmp(&right.0));
-        }
-        _ => {}
-    }
-}
-
-/// Compare sorted input values, which means specifically establishing an order between the variants
-/// of input values, and comparing values for the same variants accordingly. This is used for
-/// hashing directives in a way consistent with [same_directives()].
-///
-/// Note that Floats and Ints are compared textually and not parsed numerically. This is fine for
-/// the purposes of hashing. For object comparison semantics, see [compare_sorted_object_pairs()].
-fn compare_sorted_value(left: &executable::Value, right: &executable::Value) -> std::cmp::Ordering {
-    use apollo_compiler::executable::Value;
-    fn discriminant(value: &Value) -> u8 {
-        match value {
-            Value::Null => 0,
-            Value::Enum(_) => 1,
-            Value::Variable(_) => 2,
-            Value::String(_) => 3,
-            Value::Float(_) => 4,
-            Value::Int(_) => 5,
-            Value::Boolean(_) => 6,
-            Value::List(_) => 7,
-            Value::Object(_) => 8,
-        }
-    }
-    match (left, right) {
-        (Value::Null, Value::Null) => std::cmp::Ordering::Equal,
-        (Value::Enum(left), Value::Enum(right)) => left.cmp(right),
-        (Value::Variable(left), Value::Variable(right)) => left.cmp(right),
-        (Value::String(left), Value::String(right)) => left.cmp(right),
-        (Value::Float(left), Value::Float(right)) => left.as_str().cmp(right.as_str()),
-        (Value::Int(left), Value::Int(right)) => left.as_str().cmp(right.as_str()),
-        (Value::Boolean(left), Value::Boolean(right)) => left.cmp(right),
-        (Value::List(left), Value::List(right)) => left.len().cmp(&right.len()).then_with(|| {
-            left.iter()
-                .zip(right)
-                .map(|(left, right)| compare_sorted_value(left, right))
-                .find(|o| o.is_ne())
-                .unwrap_or(std::cmp::Ordering::Equal)
-        }),
-        (Value::Object(left), Value::Object(right)) => compare_sorted_name_value_pairs(
-            left.iter().map(|pair| &pair.0),
-            left.iter().map(|pair| &pair.1),
-            right.iter().map(|pair| &pair.0),
-            right.iter().map(|pair| &pair.1),
-        ),
-        _ => discriminant(left).cmp(&discriminant(right)),
-    }
-}
-
-/// Compare the (name, value) pair iterators, which are assumed to be sorted by name and have sorted
-/// values. This is used for hashing objects/arguments in a way consistent with [same_directives()].
-///
-/// Note that pair iterators are compared by length, then lexicographically by name, then finally
-/// recursively by value. This is intended to compute an ordering quickly for hashing.
-fn compare_sorted_name_value_pairs<'doc>(
-    left_names: impl ExactSizeIterator<Item = &'doc Name>,
-    left_values: impl ExactSizeIterator<Item = &'doc Node<executable::Value>>,
-    right_names: impl ExactSizeIterator<Item = &'doc Name>,
-    right_values: impl ExactSizeIterator<Item = &'doc Node<executable::Value>>,
-) -> std::cmp::Ordering {
-    left_names
-        .len()
-        .cmp(&right_names.len())
-        .then_with(|| left_names.cmp(right_names))
-        .then_with(|| {
-            left_values
-                .zip(right_values)
-                .map(|(left, right)| compare_sorted_value(left, right))
-                .find(|o| o.is_ne())
-                .unwrap_or(std::cmp::Ordering::Equal)
-        })
-}
-
-/// Sort arguments, which means specifically sorting arguments by names and object values by keys
-/// (assuming no duplicates). This is used for hashing arguments in a way consistent with
-/// [same_arguments()].
-pub(super) fn sort_arguments(arguments: &mut [Node<executable::Argument>]) {
-    arguments
-        .iter_mut()
-        .for_each(|arg| sort_value(arg.make_mut().value.make_mut()));
-    arguments.sort_by(|left, right| left.name.cmp(&right.name));
-}
-
-/// Compare sorted arguments; see [compare_sorted_name_value_pairs()] for semantics. This is used
-/// for hashing directives in a way consistent with [same_directives()].
-pub(super) fn compare_sorted_arguments(
-    left: &[Node<executable::Argument>],
-    right: &[Node<executable::Argument>],
-) -> std::cmp::Ordering {
-    compare_sorted_name_value_pairs(
-        left.iter().map(|arg| &arg.name),
-        left.iter().map(|arg| &arg.value),
-        right.iter().map(|arg| &arg.name),
-        right.iter().map(|arg| &arg.value),
-    )
-}
 
 pub(super) fn is_deferred_selection(directives: &executable::DirectiveList) -> bool {
     directives.has("defer")
