+++ apollo-federation/src/operation/directive_list.rs
@@ -8,10 +8,94 @@ use std::sync::Arc;
 use std::sync::OnceLock;
 
 use apollo_compiler::executable;
+use apollo_compiler::Name;
 use apollo_compiler::Node;
 
-use super::compare_sorted_arguments;
 use super::sort_arguments;
+
+/// Compare sorted input values, which means specifically establishing an order between the variants
+/// of input values, and comparing values for the same variants accordingly. This is used for
+/// hashing directives in a way consistent with [same_directives()].
+///
+/// Note that Floats and Ints are compared textually and not parsed numerically. This is fine for
+/// the purposes of hashing. For object comparison semantics, see [compare_sorted_object_pairs()].
+fn compare_sorted_value(left: &executable::Value, right: &executable::Value) -> std::cmp::Ordering {
+    use apollo_compiler::executable::Value;
+    fn discriminant(value: &Value) -> u8 {
+        match value {
+            Value::Null => 0,
+            Value::Enum(_) => 1,
+            Value::Variable(_) => 2,
+            Value::String(_) => 3,
+            Value::Float(_) => 4,
+            Value::Int(_) => 5,
+            Value::Boolean(_) => 6,
+            Value::List(_) => 7,
+            Value::Object(_) => 8,
+        }
+    }
+    match (left, right) {
+        (Value::Null, Value::Null) => std::cmp::Ordering::Equal,
+        (Value::Enum(left), Value::Enum(right)) => left.cmp(right),
+        (Value::Variable(left), Value::Variable(right)) => left.cmp(right),
+        (Value::String(left), Value::String(right)) => left.cmp(right),
+        (Value::Float(left), Value::Float(right)) => left.as_str().cmp(right.as_str()),
+        (Value::Int(left), Value::Int(right)) => left.as_str().cmp(right.as_str()),
+        (Value::Boolean(left), Value::Boolean(right)) => left.cmp(right),
+        (Value::List(left), Value::List(right)) => left.len().cmp(&right.len()).then_with(|| {
+            left.iter()
+                .zip(right)
+                .map(|(left, right)| compare_sorted_value(left, right))
+                .find(|o| o.is_ne())
+                .unwrap_or(std::cmp::Ordering::Equal)
+        }),
+        (Value::Object(left), Value::Object(right)) => compare_sorted_name_value_pairs(
+            left.iter().map(|pair| &pair.0),
+            left.iter().map(|pair| &pair.1),
+            right.iter().map(|pair| &pair.0),
+            right.iter().map(|pair| &pair.1),
+        ),
+        _ => discriminant(left).cmp(&discriminant(right)),
+    }
+}
+
+/// Compare the (name, value) pair iterators, which are assumed to be sorted by name and have sorted
+/// values. This is used for hashing objects/arguments in a way consistent with [same_directives()].
+///
+/// Note that pair iterators are compared by length, then lexicographically by name, then finally
+/// recursively by value. This is intended to compute an ordering quickly for hashing.
+fn compare_sorted_name_value_pairs<'doc>(
+    left_names: impl ExactSizeIterator<Item = &'doc Name>,
+    left_values: impl ExactSizeIterator<Item = &'doc Node<executable::Value>>,
+    right_names: impl ExactSizeIterator<Item = &'doc Name>,
+    right_values: impl ExactSizeIterator<Item = &'doc Node<executable::Value>>,
+) -> std::cmp::Ordering {
+    left_names
+        .len()
+        .cmp(&right_names.len())
+        .then_with(|| left_names.cmp(right_names))
+        .then_with(|| {
+            left_values
+                .zip(right_values)
+                .map(|(left, right)| compare_sorted_value(left, right))
+                .find(|o| o.is_ne())
+                .unwrap_or(std::cmp::Ordering::Equal)
+        })
+}
+
+/// Compare sorted arguments; see [compare_sorted_name_value_pairs()] for semantics. This is used
+/// for hashing directives in a way consistent with [same_directives()].
+fn compare_sorted_arguments(
+    left: &[Node<executable::Argument>],
+    right: &[Node<executable::Argument>],
+) -> std::cmp::Ordering {
+    compare_sorted_name_value_pairs(
+        left.iter().map(|arg| &arg.name),
+        left.iter().map(|arg| &arg.value),
+        right.iter().map(|arg| &arg.name),
+        right.iter().map(|arg| &arg.value),
+    )
+}
 
 /// An empty apollo-compiler directive list that we can return a reference to when a
 /// [`DirectiveList`] is in the empty state.
